+++ lib/src/presets/silky_grid_view.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart' show ScrollCacheExtent;
 
 import '../silky_scroll_animator.dart';
 import '../silky_scroll_config.dart';
@@ -45,6 +46,7 @@ class SilkyGridView extends StatelessWidget {
     this.addRepaintBoundaries = true,
     this.addSemanticIndexes = true,
     this.cacheExtent,
+    this.scrollCacheExtent,
     this.semanticChildCount,
     this.dragStartBehavior = DragStartBehavior.start,
     this.keyboardDismissBehavior = ScrollViewKeyboardDismissBehavior.manual,
@@ -94,6 +96,7 @@ class SilkyGridView extends StatelessWidget {
     this.addRepaintBoundaries = true,
     this.addSemanticIndexes = true,
     this.cacheExtent,
+    this.scrollCacheExtent,
     this.semanticChildCount,
     this.dragStartBehavior = DragStartBehavior.start,
     this.keyboardDismissBehavior = ScrollViewKeyboardDismissBehavior.manual,
@@ -145,6 +148,7 @@ class SilkyGridView extends StatelessWidget {
     this.addRepaintBoundaries = true,
     this.addSemanticIndexes = true,
     this.cacheExtent,
+    this.scrollCacheExtent,
     this.semanticChildCount,
     this.dragStartBehavior = DragStartBehavior.start,
     this.keyboardDismissBehavior = ScrollViewKeyboardDismissBehavior.manual,
@@ -198,6 +202,7 @@ class SilkyGridView extends StatelessWidget {
     this.addRepaintBoundaries = true,
     this.addSemanticIndexes = true,
     this.cacheExtent,
+    this.scrollCacheExtent,
     this.semanticChildCount,
     this.dragStartBehavior = DragStartBehavior.start,
     this.keyboardDismissBehavior = ScrollViewKeyboardDismissBehavior.manual,
@@ -251,6 +256,7 @@ class SilkyGridView extends StatelessWidget {
   final bool addRepaintBoundaries;
   final bool addSemanticIndexes;
   final double? cacheExtent;
+  final ScrollCacheExtent? scrollCacheExtent;
   final int? semanticChildCount;
   final DragStartBehavior dragStartBehavior;
   final ScrollViewKeyboardDismissBehavior keyboardDismissBehavior;
@@ -275,6 +281,9 @@ class SilkyGridView extends StatelessWidget {
   Widget build(BuildContext context) {
     final cfg = silkyConfig;
     final effectiveDirection = cfg?.direction ?? scrollDirection;
+    final effectiveScrollCacheExtent =
+        scrollCacheExtent ??
+        (cacheExtent == null ? null : ScrollCacheExtent.pixels(cacheExtent!));
 
     return SilkyScroll(
       controller: controller,
@@ -316,7 +325,7 @@ class SilkyGridView extends StatelessWidget {
             addAutomaticKeepAlives: addAutomaticKeepAlives,
             addRepaintBoundaries: addRepaintBoundaries,
             addSemanticIndexes: addSemanticIndexes,
-            cacheExtent: cacheExtent,
+            scrollCacheExtent: effectiveScrollCacheExtent,
             semanticChildCount: semanticChildCount,
             dragStartBehavior: dragStartBehavior,
             keyboardDismissBehavior: keyboardDismissBehavior,
@@ -335,7 +344,7 @@ class SilkyGridView extends StatelessWidget {
             addAutomaticKeepAlives: addAutomaticKeepAlives,
             addRepaintBoundaries: addRepaintBoundaries,
             addSemanticIndexes: addSemanticIndexes,
-            cacheExtent: cacheExtent,
+            scrollCacheExtent: effectiveScrollCacheExtent,
             semanticChildCount: semanticChildCount,
             dragStartBehavior: dragStartBehavior,
             keyboardDismissBehavior: keyboardDismissBehavior,
@@ -356,7 +365,7 @@ class SilkyGridView extends StatelessWidget {
             addAutomaticKeepAlives: addAutomaticKeepAlives,
             addRepaintBoundaries: addRepaintBoundaries,
             addSemanticIndexes: addSemanticIndexes,
-            cacheExtent: cacheExtent,
+            scrollCacheExtent: effectiveScrollCacheExtent,
             semanticChildCount: semanticChildCount,
             dragStartBehavior: dragStartBehavior,
             keyboardDismissBehavior: keyboardDismissBehavior,
@@ -378,7 +387,7 @@ class SilkyGridView extends StatelessWidget {
             addAutomaticKeepAlives: addAutomaticKeepAlives,
             addRepaintBoundaries: addRepaintBoundaries,
             addSemanticIndexes: addSemanticIndexes,
-            cacheExtent: cacheExtent,
+            scrollCacheExtent: effectiveScrollCacheExtent,
             semanticChildCount: semanticChildCount,
             dragStartBehavior: dragStartBehavior,
             keyboardDismissBehavior: keyboardDismissBehavior,